requirements.txt
# Required for loading pre-computed embeddings (no API keys needed)
numpy
pandas
pyarrow

# Required only for generating new embeddings
openai        # LLM enrichment (Step 1)
google-genai  # Gemini embeddings (Step 2)
pydantic
tqdm
